move logging test to no race
