+++ rest/changestest/changes_api_test.go
@@ -4399,27 +4399,3 @@ func WriteDirectWithKey(t *testing.T, key string, channelArray []string, sequenc
 	require.NoError(t, err)
 
 }
-
-// TestDocChangedLogging exercises some of the logging in DocChanged
-func TestDocChangedLogging(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyHTTP, base.KeyCache, base.KeyChanges)
-
-	rt := rest.NewRestTesterMultipleCollections(t, nil, 2)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/{{.keyspace1}}/doc1", `{"foo":"bar"}`)
-	rest.RequireStatus(t, response, http.StatusCreated)
-	response = rt.SendAdminRequest("PUT", "/{{.keyspace2}}/doc1", `{"foo":"bar"}`)
-	rest.RequireStatus(t, response, http.StatusCreated)
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	base.AssertLogContains(t, "Ignoring non-metadata mutation for doc", func() {
-		err := rt.GetDatabase().MetadataStore.Set("doc1", 0, nil, db.Body{"foo": "bar"})
-		require.NoError(t, err)
-		// write another doc to ensure the previous non-metadata doc has been seen...
-		// no other way of synchronising this no-op as no stats to wait on
-		response = rt.SendAdminRequest("PUT", "/{{.keyspace1}}/doc2", `{"foo":"bar"}`)
-		rest.RequireStatus(t, response, http.StatusCreated)
-		require.NoError(t, rt.WaitForPendingChanges())
-	})
-}
